+++ src/onegov/core/cli/commands.py
@@ -135,15 +135,14 @@ def sendmail(group_context, queue, limit):
               help="Do not transfer the files")
 @click.option('--no-database', default=False, is_flag=True,
               help="Do not transfer the database")
-@click.option('--transfer-schema', help="Only transfer this schema")
-@click.option('--to-14', default=False, is_flag=True,
-              help="Transfer from postgres <14 to >=14.")
+@click.option('--transfer-schema',
+              help="Only transfer this schema, e.g. /town6/govikon")
 @click.option('--add-admins', default=False, is_flag=True,
-              help="Add local admins")
+              help="Add local admins (admin@example.org:test)")
 @pass_group_context
 def transfer(group_context,
              server, remote_config, confirm, no_filestorage, no_database,
-             transfer_schema, to_14, add_admins):
+             transfer_schema, add_admins):
     """ Transfers the database and all files from a server running a
     onegov-cloud application and installs them locally, overwriting the
     local data!
@@ -231,6 +230,10 @@ def transfer(group_context,
         schemas = (s for s in schemas if fnmatch(s, schema_glob))
         schemas = tuple(schemas)
 
+        if not schemas:
+            click.echo("No matching schema(s) found!")
+            return schemas
+
         # Prepare send command
         send = f"ssh {server} sudo -u postgres nice -n 10 pg_dump {remote_db}"
         send = f"{send} --no-owner --no-privileges"
@@ -256,9 +259,6 @@ def transfer(group_context,
 
         # Transfer
         click.echo("Transfering database")
-        if to_14:
-            # todo: remove me after 2022.43 has been deployed everywhere
-            recv = f"sed 's/anyarray/anycompatiblearray/g' | {recv}"
         if shutil.which('pv'):
             recv = f'pv --name "{remote_db}@postgres" -r -b | {recv}'
         subprocess.check_output(f'{send} | {recv}', shell=True)
@@ -348,12 +348,12 @@ def transfer(group_context,
 
         click.echo(f"Fetching {remote_cfg.namespace}")
 
+        if not no_database:
+            schemas.update(transfer_database_of_app(local_cfg, remote_cfg))
+
         if not no_filestorage:
             transfer_storage_of_app(local_cfg, remote_cfg)
             transfer_depot_storage_of_app(local_cfg, remote_cfg)
-
-        if not no_database:
-            schemas.update(transfer_database_of_app(local_cfg, remote_cfg))
 
     if add_admins:
         for schema in schemas: