+++ src/onegov/org/forms/reservation.py
@@ -8,9 +8,8 @@ from wtforms.fields import DateField
 from wtforms.fields import EmailField
 from wtforms.fields import RadioField
 from wtforms.validators import DataRequired, Email, InputRequired
-
-from .allocation import WEEKDAYS
-
+from onegov.core.csv import convert_list_of_list_of_dicts_to_xlsx
+from onegov.org.forms.allocation import WEEKDAYS
 
 # include all fields used below so we can filter them out
 # when we merge this form with the custom form definition
@@ -18,7 +17,6 @@ RESERVED_FIELDS = ['email']
 
 
 class ReservationForm(Form):
-
     reserved_fields = RESERVED_FIELDS
 
     email = EmailField(
@@ -28,7 +26,6 @@ class ReservationForm(Form):
 
 
 class FindYourSpotForm(Form):
-
     start = DateField(
         label=_("From"),
         validators=[InputRequired()])
@@ -167,3 +164,18 @@ class FindYourSpotForm(Form):
             if start <= date <= end:
                 return True
         return False
+
+
+class ExportToExcelWorksheets(Form):
+    """ A form providing the export of multiple reservations into Worksheets
+    """
+
+    @property
+    def format(self):
+        return 'xlsx'
+
+    def as_multiple_export_response(self, keys, results, titles):
+
+        return convert_list_of_list_of_dicts_to_xlsx(results,
+                                                     titles_list=titles,
+                                                     key_list=keys)