+++ src/onegov/org/views/reservation.py
@@ -8,6 +8,7 @@ from dill import pickles
 from libres.modules.errors import LibresError
 from onegov.core.custom import json
 from onegov.core.security import Public, Private
+from onegov.core.templates import render_template
 from onegov.form import FormCollection, merge_forms, as_internal_id
 from onegov.org import _, OrgApp
 from onegov.org import utils
@@ -15,9 +16,11 @@ from onegov.org.cli import close_ticket
 from onegov.org.elements import Link
 from onegov.org.forms import ReservationForm, InternalTicketChatMessageForm
 from onegov.org.layout import ReservationLayout, TicketChatMessageLayout
+from onegov.org.layout import DefaultMailLayout
 from onegov.org.mail import send_ticket_mail
 from onegov.org.models import (
-    TicketMessage, TicketChatMessage, ReservationMessage)
+    TicketMessage, TicketChatMessage, ReservationMessage,
+    ResourceRecipient, ResourceRecipientCollection)
 from onegov.org.models.resource import FindYourSpotCollection
 from onegov.reservation import Allocation, Reservation, Resource
 from onegov.ticket import TicketCollection
@@ -637,6 +640,49 @@ def accept_reservation(self, request, text=None, notify=False):
                 'message': message
             }
         )
+
+        # get all recipients which require an e-mail for this resource
+        q = ResourceRecipientCollection(request.session).query()
+        q = q.filter(ResourceRecipient.medium == 'email')
+        q = q.order_by(None).order_by(ResourceRecipient.address)
+        q = q.with_entities(
+            ResourceRecipient.address,
+            ResourceRecipient.content
+        )
+        recipients = [
+            r.address
+            for r in q if (
+                self.resource.hex in r.content['resources']
+                and r.content.get('new_reservations', False)
+            )
+        ]
+
+        # E-mail for new reservations
+        args = {
+            'layout': DefaultMailLayout(object(), request),
+            'title': request.translate(
+                _("${org} New Reservation(s)", mapping={
+                    'org': request.app.org.title
+                })
+            ),
+            'form': form,
+            'model': self,
+            'resource': resource,
+            'reservations': reservations,
+            'show_submission': show_submission,
+            'message': message
+        }
+
+        content = render_template(
+            'mail_new_reservation_notification.pt', request, args
+        )
+
+        for r in recipients:
+            request.app.send_transactional_email(
+                subject=args['title'],
+                receivers=(r),
+                content=content,
+            )
 
         request.success(_("The reservations were accepted"))
     else: