+++ tests/onegov/org/test_views_settings.py
@@ -43,16 +43,6 @@ def test_settings(client):
     assert '<img src="https://seantis.ch/logo.img"' in settings.text
     assert '<style>h1 { text-decoration: underline; }</style>' in settings.text
 
-    # homepage settings
-    settings = client.get('/homepage-settings')
-    settings.form['homepage_image_1'] = "http://images/one"
-    settings.form['homepage_image_2'] = "http://images/two"
-    settings.form.submit()
-
-    settings = client.get('/homepage-settings')
-    assert 'http://images/one' in settings
-    assert 'http://images/two' in settings
-
     # analytics settings
     settings = client.get('/analytics-settings')
     settings.form['analytics_code'] = '<script>alert("Hi!");</script>'
@@ -69,10 +59,10 @@ def test_settings(client):
 
     # test default not giving the color
     assert settings.form['left_header_color'].value == '#000000'
-    assert settings.form['left_header_announcement_bg_color'].value == (
+    assert settings.form['announcement_bg_color'].value == (
         '#FBBC05'
     )
-    assert settings.form['left_header_announcement_font_color'].value == (
+    assert settings.form['announcement_font_color'].value == (
         '#000000'
     )
 
@@ -80,18 +70,23 @@ def test_settings(client):
     settings.form['left_header_url'] = 'https://govikon.ch'
     settings.form['left_header_rem'] = 2.5
     settings.form['left_header_color'] = color
-    settings.form['left_header_announcement'] = text
-    settings.form['left_header_announcement_bg_color'] = bg_color
-    settings.form['left_header_announcement_font_color'] = color
+    settings.form['announcement'] = text
+    settings.form['announcement_url'] = 'https://other-town.ch'
+    settings.form['announcement_bg_color'] = bg_color
+    settings.form['announcement_font_color'] = color
     page = settings.form.submit().follow()
     assert (
         f'<a href="https://govikon.ch" '
         f'style="color:{color}; font-size: 2.5rem">'
     ) in page
     assert text in page
+    assert '' in page
     assert (
-        f'<div id="announcement" style="color: {color}; '
-        f'background-color: {bg_color};">'
+        f'<div id="announcement_header" '
+        f'style="background-color: {bg_color};">'
+    ) in page
+    assert (
+        f'<a style="color: {color}" href="https://other-town.ch"'
     ) in page